analysis/requirement.txt
kafka-python
psycopg2-binary
pyspark
